requirements.txt
# Core Django
Django>=5.0,<5.1
djangorestframework>=3.15.0
django-filter>=24.2  # para filtros en DRF (DjangoFilterBackend)

# CORS para conectar con Vue u otro frontend
django-cors-headers>=4.3.0

# Variables de entorno (.env)
python-dotenv>=1.0.0

# Tareas asíncronas (Celery + Redis como broker/result backend)
celery>=5.3.0
redis>=5.0.0

# Servidor WSGI (opcional, útil si algún día despliegas en un server)
gunicorn>=21.2.0

# Envío de correos (vía API / SMTP)
sendgrid>=6.11.0

# Imágenes (ImageField, avatares, etc.)
Pillow>=10.0.0

# Documentación OpenAPI / Swagger
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2024.3.1
